+++ packages/marketplace-management/src/services/user.ts
@@ -10,19 +10,22 @@ interface UpdateUserGroupParams {
 
 export const addMemberToGroup = async (group: UpdateUserGroupParams): Promise<any | undefined> => {
   try {
-    const response = await fetcher({
-      api: window.reapit.config.platformApiUrl,
-      url: `${URLS.USERS_GROUPS}/${group.id}/members`,
-      method: 'POST',
-      headers: await getPlatformHeaders(reapitConnectBrowserSession),
-      body: group,
-    })
+    const headers = await getPlatformHeaders(reapitConnectBrowserSession)
+    if (headers) {
+      const response = await fetcher({
+        api: window.reapit.config.platformApiUrl,
+        url: `${URLS.USERS_GROUPS}/${group.id}/members`,
+        method: 'POST',
+        headers,
+        body: group,
+      })
 
-    if (response) {
-      return response
+      if (response) {
+        return response
+      }
+
+      throw new Error('Adding member to group failed')
     }
-
-    throw new Error('Adding member to group failed')
   } catch (err) {
     logger(err)
   }
@@ -30,19 +33,22 @@ export const addMemberToGroup = async (group: UpdateUserGroupParams): Promise<an
 
 export const removeMemberFromGroup = async (group: UpdateUserGroupParams): Promise<any | undefined> => {
   try {
-    const response = await fetcher({
-      api: window.reapit.config.platformApiUrl,
-      url: `${URLS.USERS_GROUPS}/${group.id}/members/${group.userId}`,
-      method: 'DELETE',
-      headers: await getPlatformHeaders(reapitConnectBrowserSession),
-      body: group,
-    })
+    const headers = await getPlatformHeaders(reapitConnectBrowserSession, 'latest')
+    if (headers) {
+      const response = await fetcher({
+        api: window.reapit.config.platformApiUrl,
+        url: `${URLS.USERS_GROUPS}/${group.id}/members/${group.userId}`,
+        method: 'DELETE',
+        headers,
+        body: group,
+      })
 
-    if (response) {
-      return response
+      if (response) {
+        return response
+      }
+
+      throw new Error('Removing member from group failed')
     }
-
-    throw new Error('Removing member from group failed')
   } catch (err) {
     logger(err)
   }
